% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{execute_dataset_joins}
\alias{execute_dataset_joins}
\title{Voer joins uit om datasets te combineren op basis van een verbindingspad}
\usage{
execute_dataset_joins(tibbles, path, verbose, reference_date)
}
\arguments{
\item{tibbles}{Lijst met tibbles om te combineren}

\item{path}{Verbindingspad dat specificeert hoe de tibbles moeten worden gekoppeld}

\item{verbose}{Logische waarde die aangeeft of gedetailleerde verbindingsinformatie moet worden weergegeven}

\item{reference_date}{Optionele datum om geldige records te filteren}
}
\value{
Een gecombineerde tibble
}
\description{
Voer joins uit om datasets te combineren op basis van een verbindingspad
}
\keyword{internal}
