% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{find_best_connection_path}
\alias{find_best_connection_path}
\title{Vind het beste verbindingspad uit een lijst met paden}
\usage{
find_best_connection_path(paths)
}
\arguments{
\item{paths}{Een lijst met verbindingspaden zoals geretourneerd door rio_find_connection_paths()}
}
\value{
Het beste verbindingspad op basis van lengte en betrouwbaarheid
}
\description{
Vind het beste verbindingspad uit een lijst met paden
}
\keyword{internal}
