% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{find_candidate_joining_fields}
\alias{find_candidate_joining_fields}
\title{Find candidate joining fields between two datasets}
\usage{
find_candidate_joining_fields(df1, df2, name1, name2)
}
\arguments{
\item{df1}{First dataframe or tibble}

\item{df2}{Second dataframe or tibble}

\item{name1}{Name of the first dataset}

\item{name2}{Name of the second dataset}
}
\value{
A list with details about the potential connection
}
\description{
Find candidate joining fields between two datasets
}
\keyword{internal}
