% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{find_date_fields}
\alias{find_date_fields}
\title{Find date fields in a dataframe}
\usage{
find_date_fields(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A character vector of date field names
}
\description{
Find date fields in a dataframe
}
\keyword{internal}
