% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{get_resource_id_from_name}
\alias{get_resource_id_from_name}
\title{Helper function to get resource ID from name}
\usage{
get_resource_id_from_name(
  conn = NULL,
  resource_name,
  package_id = "rio_nfo_po_vo_vavo_mbo_ho"
)
}
\arguments{
\item{conn}{Connection object}

\item{resource_name}{Resource name}

\item{package_id}{Package ID}
}
\value{
Resource ID or NULL if not found
}
\description{
Helper function to get resource ID from name
}
\keyword{internal}
