% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{get_rio_data_dir}
\alias{get_rio_data_dir}
\title{Verkrijg de directory voor het opslaan van RIO package data}
\usage{
get_rio_data_dir()
}
\value{
Een character string met het pad naar de RIO data directory
}
\description{
Verkrijg de directory voor het opslaan van RIO package data
}
\keyword{internal}
