% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{nest_result}
\alias{nest_result}
\title{Create a nested tibble from join results}
\usage{
nest_result(joined_data, dataset_names, path_info)
}
\arguments{
\item{joined_data}{The result of joining datasets}

\item{dataset_names}{Names of the original datasets}

\item{path_info}{Path information as returned by find_optimal_path()}
}
\value{
A nested tibble
}
\description{
This function reorganizes the joined data into a nested tibble structure,
which can be more intuitive for certain types of analysis.
}
\keyword{internal}
