% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{rio_add_relation}
\alias{rio_add_relation}
\title{Add a relation to a list of RIO relations}
\usage{
rio_add_relation(relations, relation)
}
\arguments{
\item{relations}{The list of RIO relations.}

\item{relation}{The relation to add or a list with arguments for rio_create_relation().}
}
\value{
The updated list of RIO relations.
}
\description{
This function adds a relation to a list of RIO relations.
}
\keyword{internal}
