% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{rio_api_call}
\alias{rio_api_call}
\title{Execute an API call to the RIO CKAN API}
\usage{
rio_api_call(
  req = NULL,
  endpoint,
  method = "GET",
  body = NULL,
  simplify = TRUE
)
}
\arguments{
\item{req}{A request object, typically created with \code{rio_api_connection()}.
If NULL, a new connection will be created.}

\item{endpoint}{The API endpoint to call.}

\item{method}{The HTTP method to use. Default is "GET".}

\item{body}{The request body as a list for POST requests. Default is NULL.}

\item{simplify}{Whether to simplify the response to R vectors/matrices. Default is TRUE.}
}
\value{
The parsed JSON response from the API.
}
\description{
This function executes a prepared request to the RIO CKAN API and returns
the parsed JSON response.
}
\examples{
\dontrun{
response <- rio_api_call(endpoint = "group_list")
}

}
\keyword{internal}
