% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{rio_api_check}
\alias{rio_api_check}
\title{Check API availability}
\usage{
rio_api_check()
}
\value{
TRUE if the API is available, FALSE otherwise.
}
\description{
This function checks if the RIO CKAN API is available and responsive.
}
\examples{
\dontrun{
is_available <- rio_api_check()
}

}
\keyword{internal}
