% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_combine}
\alias{rio_combine}
\title{Combine RIO datasets}
\usage{
rio_combine(..., by = NULL, suffix = c("_x", "_y"))
}
\arguments{
\item{...}{Named tibbles to combine. Each argument should be a tibble with data from RIO.}

\item{by}{Character vector of columns to join by. If NULL, will join by all columns with the same name.}

\item{suffix}{Suffixes to add to make non-by column names unique across the joined tables.}
}
\value{
A combined tibble with data from all input tibbles.
}
\description{
This function combines multiple RIO datasets based on a common ID field.
}
\examples{
\dontrun{
rio_conn <- rio_api_connection()
locations <- rio_get_locations(rio_conn, city = "Rotterdam")
institutions <- rio_fetch_data(rio_conn, resource_id = "institution-resource-id")

# Combine datasets
combined_data <- rio_combine(
  locations = locations,
  institutions = institutions,
  by = "INSTELLINGSCODE"
)
}

}
