% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_combine_smart}
\alias{rio_combine_smart}
\title{Intelligente koppeling van datasets op basis van gedetecteerde structuur}
\usage{
rio_combine_smart(
  ...,
  auto_detect = TRUE,
  max_path_length = 3,
  reference_date = NULL,
  verbose = TRUE
)
}
\arguments{
\item{...}{Benoemde tibbles om te koppelen}

\item{auto_detect}{Logische waarde die aangeeft of automatisch een structuur moet worden gedetecteerd als er geen bestaat}

\item{max_path_length}{Maximale padlengte tussen datasets}

\item{reference_date}{Optionele datum om geldige records te filteren}

\item{verbose}{Logische waarde die aangeeft of gedetailleerde informatie moet worden weergegeven}
}
\value{
Een gekoppelde tibble
}
\description{
Intelligente koppeling van datasets op basis van gedetecteerde structuur
}
\keyword{internal}
