% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{rio_export_relations_csv}
\alias{rio_export_relations_csv}
\title{Export RIO relations to a CSV file}
\usage{
rio_export_relations_csv(relations, file_path = "rio_relations.csv")
}
\arguments{
\item{relations}{The list of RIO relations.}

\item{file_path}{Path to the CSV file where the relations will be saved.}
}
\value{
Invisibly the name of the file where the relations were saved.
}
\description{
This function exports RIO relations to a CSV file.
}
\keyword{internal}
