% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{rio_fetch_data}
\alias{rio_fetch_data}
\title{Fetch data from a RIO dataset}
\usage{
rio_fetch_data(
  conn = NULL,
  resource_id = NULL,
  resource_name = NULL,
  limit = NULL,
  query = NULL,
  package_id = "rio_nfo_po_vo_vavo_mbo_ho",
  all_records = TRUE,
  batch_size = 1000,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{conn}{A connection object created with \code{rio_api_connection()}.
If NULL, a new connection will be created.}

\item{resource_id}{The ID of the dataset resource to retrieve. Default is NULL.}

\item{resource_name}{The name of the dataset resource to retrieve. Default is NULL.
Either resource_id or resource_name must be provided.}

\item{limit}{Maximum number of records to return. Default is NULL (all records).
Specify a number to limit the results.}

\item{query}{A search query string for full-text search. Default is NULL.}

\item{package_id}{The ID of the package to search in when using resource_name.
Default is "rio_nfo_po_vo_vavo_mbo_ho".}

\item{all_records}{Logical indicating whether to fetch all records (which may
require multiple API calls). Default is TRUE.}

\item{batch_size}{Number of records to retrieve per API call. Default is 1000,
which is the maximum supported by the API.}

\item{quiet}{Logical indicating whether to suppress progress messages. Default is FALSE.}

\item{...}{Additional named parameters used as filters. For example, PLAATSNAAM = "Rotterdam"
will filter for records where the PLAATSNAAM field equals "Rotterdam".}
}
\value{
A tibble containing the retrieved data.
}
\description{
This function retrieves data from a specific dataset in the RIO CKAN API.
You can specify either a resource_id or a resource_name to identify the dataset.
Additional filter parameters can be provided to filter the data.
By default, this function retrieves all available records (not just the first 1000).
}
\examples{
\dontrun{
# Get all educational locations in Rotterdam
locations <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  PLAATSNAAM = "Rotterdam"
)

# Get only the first 500 records
limited_data <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  limit = 500
)

# Get the first 1000 records without attempting to fetch all records
first_batch <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  all_records = FALSE
)

# Fetch data without showing progress bar
silent_fetch <- rio_fetch_data(
  resource_name = "Onderwijslocaties",
  quiet = TRUE
)
}

}
