% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{rio_filter_by_date}
\alias{rio_filter_by_date}
\title{Filter data based on reference date}
\usage{
rio_filter_by_date(
  data,
  reference_date = Sys.Date(),
  start_date_col = "BEGINDATUM",
  end_date_col = "EINDDATUM"
)
}
\arguments{
\item{data}{A tibble containing RIO data.}

\item{reference_date}{The reference date to filter by. Default is the current date.}

\item{start_date_col}{The name of the column containing start dates. Default is "BEGINDATUM".}

\item{end_date_col}{The name of the column containing end dates. Default is "EINDDATUM".}
}
\value{
A tibble containing only records that are valid at the reference date.
}
\description{
This function filters data from the RIO API based on a reference date,
considering start and end dates of records.
}
\examples{
\dontrun{
# Filter data for records valid on January 1, 2023
filtered_data <- rio_filter_by_date(
  data,
  reference_date = as.Date("2023-01-01")
)
}

}
