% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{rio_filter_by_reference_date}
\alias{rio_filter_by_reference_date}
\title{Filter data based on a reference date}
\usage{
rio_filter_by_reference_date(
  data,
  reference_date,
  start_date_col = NULL,
  end_date_col = NULL
)
}
\arguments{
\item{data}{A tibble containing temporal data with date columns.}

\item{reference_date}{The date for which records should be valid (as Date object).}

\item{start_date_col}{The name of the column containing start dates.
Default is NULL, which means the function will automatically search for
appropriate columns.}

\item{end_date_col}{The name of the column containing end dates.
Default is NULL, which means the function will automatically search for
appropriate columns.}
}
\value{
A filtered tibble containing only records that are valid at the reference date,
or the original tibble if no suitable date columns are found.
}
\description{
This function filters temporal data based on a reference date,
keeping only records that are valid at the specified date.
It prioritizes EINDDATUM/BEGINDATUM over EINDDATUM_PERIODE/BEGINDATUM_PERIODE.
}
\keyword{internal}
