% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{rio_get_fields}
\alias{rio_get_fields}
\title{Get information about fields in a dataset}
\usage{
rio_get_fields(resource_id = NULL, resource_name = NULL)
}
\arguments{
\item{resource_id}{The ID of the dataset resource. Default is NULL.}

\item{resource_name}{The name of the dataset resource. Default is NULL.
Either resource_id or resource_name must be provided.}
}
\value{
A tibble with information about the fields in the dataset.
}
\description{
This function retrieves information about the available fields (columns) in a dataset,
which can be used for filtering in the rio_get_data function.
}
\examples{
\dontrun{
# Get fields for the "Onderwijslocaties" dataset
fields <- rio_get_fields(resource_name = "Onderwijslocaties")

# Print the field names that can be used for filtering
print(fields$id)
}

}
