% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{rio_get_metadata}
\alias{rio_get_metadata}
\title{Get metadata for the RIO dataset}
\usage{
rio_get_metadata(conn = NULL, package_id = "rio_nfo_po_vo_vavo_mbo_ho")
}
\arguments{
\item{conn}{A connection object created with \code{rio_api_connection()}.
If NULL, a new connection will be created.}

\item{package_id}{The ID of the RIO package. Default is "rio_nfo_po_vo_vavo_mbo_ho".}
}
\value{
A list containing the package metadata.
}
\description{
This function retrieves detailed metadata for the Dutch Register of Institutions
and Programs (RIO) dataset.
}
\examples{
\dontrun{
rio_metadata <- rio_get_metadata()
}

}
