% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{rio_get_most_recent}
\alias{rio_get_most_recent}
\title{Get the most recent data version}
\usage{
rio_get_most_recent(
  data,
  id_col,
  start_date_col = "BEGINDATUM",
  end_date_col = "EINDDATUM"
)
}
\arguments{
\item{data}{A tibble containing RIO data.}

\item{id_col}{The name of the column containing entity IDs.}

\item{start_date_col}{The name of the column containing start dates. Default is "BEGINDATUM".}

\item{end_date_col}{The name of the column containing end dates. Default is "EINDDATUM".}
}
\value{
A tibble containing only the most recent version of each entity.
}
\description{
This function retrieves the most recent version of each entity in the data,
based on start and end dates.
}
\examples{
\dontrun{
# Get the most recent version of each institution
recent_data <- rio_get_most_recent(
  data,
  id_col = "INSTELLINGSCODE"
)
}

}
