% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R, R/metadata.R
\name{rio_get_resource_info}
\alias{rio_get_resource_info}
\title{Get detailed information about a specific dataset}
\usage{
rio_get_resource_info(
  conn = NULL,
  resource_id = NULL,
  resource_name = NULL,
  package_id = "rio_nfo_po_vo_vavo_mbo_ho"
)

rio_get_resource_info(
  conn = NULL,
  resource_id = NULL,
  resource_name = NULL,
  package_id = "rio_nfo_po_vo_vavo_mbo_ho"
)
}
\arguments{
\item{conn}{A connection object created with \code{rio_api_connection()}.
If NULL, a new connection will be created.}

\item{resource_id}{The ID of the resource to retrieve. Default is NULL.}

\item{resource_name}{The name of the dataset resource to retrieve. Default is NULL.
Either resource_id or resource_name must be provided.}

\item{package_id}{The ID of the package to search in when using resource_name.
Default is "rio_nfo_po_vo_vavo_mbo_ho".}
}
\value{
A list containing detailed information about the dataset.

A list containing detailed information about the dataset.
}
\description{
This function retrieves detailed information about a specific dataset (resource) in the RIO package.

This function retrieves detailed information about a specific dataset (resource) in the RIO package.
}
\examples{
\dontrun{
# Get resource info using ID
resource_info_by_id <- rio_get_resource_info(resource_id = "a7e3f323-6e46-4dca-a834-369d9d520aa8")

# Get resource info using name
resource_info_by_name <- rio_get_resource_info(resource_name = "Onderwijslocaties")
}

\dontrun{
# Get resource info using ID
resource_info_by_id <- rio_get_resource_info(resource_id = "a7e3f323-6e46-4dca-a834-369d9d520aa8")

# Get resource info using name
resource_info_by_name <- rio_get_resource_info(resource_name = "Onderwijslocaties")
}

}
