% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{rio_import_relations_csv}
\alias{rio_import_relations_csv}
\title{Import relations from a CSV file}
\usage{
rio_import_relations_csv(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file with the relations.}
}
\value{
A list of RIO relations.
}
\description{
This function imports relations from a CSV file and returns a list of relations.
}
\keyword{internal}
