% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_load_structure}
\alias{rio_load_structure}
\title{Load the RIO structure mapping}
\usage{
rio_load_structure(auto_create = TRUE, days_threshold = 28, quiet = FALSE)
}
\arguments{
\item{auto_create}{Logical indicating whether to automatically create a mapping if
none exists or if it's outdated. Default is TRUE.}

\item{days_threshold}{Number of days after which a mapping is considered outdated.
Default is 28 (4 weeks).}

\item{quiet}{Logical indicating whether to suppress progress messages. Default is FALSE.}
}
\value{
A list containing the RIO structure mapping.
}
\description{
This function loads the saved RIO structure mapping. If no mapping exists or it's
outdated, it will optionally create a new one.
}
\examples{
\dontrun{
# Load the RIO structure mapping
mapping <- rio_load_structure()
}

}
