% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{rio_map_locations}
\alias{rio_map_locations}
\title{Create a map of educational locations}
\usage{
rio_map_locations(locations, popup_vars = NULL, color = "blue", radius = 5)
}
\arguments{
\item{locations}{A tibble containing educational location data from the RIO API.}

\item{popup_vars}{Character vector of variable names to include in popups. Default is NULL.}

\item{color}{The color of the markers. Default is "blue".}

\item{radius}{The radius of the markers. Default is 5.}
}
\value{
A leaflet map object.
}
\description{
This function creates a leaflet map of educational locations from RIO data.
}
\examples{
\dontrun{
locations <- rio_get_locations(city = "Rotterdam")
map <- rio_map_locations(
  locations,
  popup_vars = c("ONDERWIJSLOCATIENAAM", "PLAATSNAAM")
)
}

}
