% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{rio_remove_relation}
\alias{rio_remove_relation}
\title{Remove a relation from a list of RIO relations}
\usage{
rio_remove_relation(relations, from, to)
}
\arguments{
\item{relations}{The list of RIO relations.}

\item{from}{Name of the source dataset.}

\item{to}{Name of the target dataset.}
}
\value{
The updated list of RIO relations.
}
\description{
This function removes a relation from a list of RIO relations.
}
\keyword{internal}
