% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{rio_save_relations}
\alias{rio_save_relations}
\title{Save the RIO relations as a YAML file}
\usage{
rio_save_relations(
  relations,
  file_path = "inst/extdata/rio_relations.yaml",
  create_dir = TRUE
)
}
\arguments{
\item{relations}{The RIO relations.}

\item{file_path}{Path to the YAML file where the relations will be saved.
Default is 'inst/extdata/rio_relations.yaml'.}

\item{create_dir}{Logical indicating whether to create the directory if it doesn't exist.
Default is TRUE.}
}
\value{
Invisibly the name of the file where the relations were saved.
}
\description{
This function saves the RIO relations as a YAML file.
}
\keyword{internal}
