% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{rio_timeline}
\alias{rio_timeline}
\title{Plot timeline of institutions or entities}
\usage{
rio_timeline(
  data,
  id_col,
  name_col = NULL,
  start_date_col = "BEGINDATUM",
  end_date_col = "EINDDATUM"
)
}
\arguments{
\item{data}{A tibble containing data with start and end dates.}

\item{id_col}{The name of the column containing entity IDs.}

\item{name_col}{The name of the column containing entity names. Default is NULL.}

\item{start_date_col}{The name of the column containing start dates. Default is "BEGINDATUM".}

\item{end_date_col}{The name of the column containing end dates. Default is "EINDDATUM".}
}
\value{
A ggplot object showing the timeline.
}
\description{
This function creates a timeline plot showing the validity periods of institutions or other entities.
}
\examples{
\dontrun{
institutions <- rio_get_data(dataset_id = "institution-resource-id")
plot <- rio_timeline(
  institutions,
  id_col = "INSTELLINGSCODE",
  name_col = "INSTELLINGSNAAM"
)
}

}
