% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{rio_visualize_connections}
\alias{rio_visualize_connections}
\title{Visualize connections between RIO datasets}
\usage{
rio_visualize_connections(
  connections,
  min_confidence = "medium",
  interactive = TRUE
)
}
\arguments{
\item{connections}{List of connections between datasets}

\item{min_confidence}{Minimum confidence level for relationships to display}

\item{interactive}{Logical indicating whether to create an interactive visualization}
}
\value{
A visualization object (igraph plot or visNetwork object)
}
\description{
This function creates a visualization of connections between RIO datasets.
}
\keyword{internal}
