% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_datasets.R
\name{visualize_path}
\alias{visualize_path}
\title{Visualize the connection path between datasets}
\usage{
visualize_path(path_info, relations, dataset_names, interactive = TRUE)
}
\arguments{
\item{path_info}{Path information as returned by find_optimal_path()}

\item{relations}{List of relation definitions}

\item{dataset_names}{Character vector with names of datasets to visualize}

\item{interactive}{Logical indicating whether to create an interactive visualization}
}
\value{
A visualization object (igraph plot or visNetwork object)
}
\description{
This function creates a visualization of how datasets are connected based
on the optimal path determined by find_optimal_path().
}
\keyword{internal}
