% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{execute_dataset_joins}
\alias{execute_dataset_joins}
\title{Execute joins to combine datasets based on a connection path}
\usage{
execute_dataset_joins(tibbles, path, verbose, reference_date)
}
\arguments{
\item{tibbles}{List of tibbles to combine}

\item{path}{Connection path specifying how to join the tibbles}

\item{verbose}{Logical indicating whether to show detailed connection information}

\item{reference_date}{Optional date to filter valid records}
}
\value{
A combined tibble
}
\description{
Execute joins to combine datasets based on a connection path
}
\keyword{internal}
