% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{execute_dataset_joins}
\alias{execute_dataset_joins}
\title{Execute joins between datasets based on a connection path}
\usage{
execute_dataset_joins(
  tibbles,
  path_info,
  relations,
  verbose = TRUE,
  reference_date = NULL,
  start_dataset = NULL
)
}
\arguments{
\item{tibbles}{List of tibbles to join}

\item{path_info}{Path information as returned by find_optimal_path()}

\item{relations}{List of relation definitions from the YAML configuration}

\item{verbose}{Logical indicating whether to display progress messages}

\item{reference_date}{Optional date to filter records by}

\item{start_dataset}{Optional name of the dataset to start with. Default is NULL,
which will use the first dataset from the first connection.}
}
\value{
A joined tibble
}
\description{
This function performs joins between datasets following the optimal path
determined by find_optimal_path(), using relations defined in YAML.
}
\keyword{internal}
