% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{find_best_connection_path}
\alias{find_best_connection_path}
\title{Find the best connection path from a list of paths}
\usage{
find_best_connection_path(paths)
}
\arguments{
\item{paths}{A list of connection paths as returned by rio_find_connection_paths()}
}
\value{
The best connection path based on length and confidence
}
\description{
Find the best connection path from a list of paths
}
\keyword{internal}
