% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{find_date_fields}
\alias{find_date_fields}
\title{Vind datumvelden in een dataframe}
\usage{
find_date_fields(df)
}
\arguments{
\item{df}{Een dataframe}
}
\value{
Een character vector met namen van datumvelden
}
\description{
Vind datumvelden in een dataframe
}
\keyword{internal}
