% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connection.R
\name{get_dataset_id_from_name}
\alias{get_dataset_id_from_name}
\title{Helper function to get resource ID from name}
\usage{
get_dataset_id_from_name(dataset_name)
}
\arguments{
\item{dataset_name}{Resource name}

\item{conn}{Connection object}

\item{package_id}{Package ID}
}
\value{
Resource ID or NULL if not found
}
\description{
Helper function to get resource ID from name
}
\keyword{internal}
