% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{get_rio_data_dir}
\alias{get_rio_data_dir}
\title{Get the directory for saving RIO package data}
\usage{
get_rio_data_dir()
}
\value{
A character string with the path to the RIO data directory
}
\description{
Get the directory for saving RIO package data
}
\keyword{internal}
