% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{rio_filter_by_reference_date}
\alias{rio_filter_by_reference_date}
\title{Filter data based on a reference date}
\usage{
rio_filter_by_reference_date(
  data,
  reference_date,
  start_date_col = NULL,
  end_date_col = NULL
)
}
\arguments{
\item{data}{A tibble containing temporal data with date columns.}

\item{reference_date}{The date for which records should be valid (as Date object).}

\item{start_date_col}{The name of the column containing start dates.
Default is NULL, which means the function will automatically search for
appropriate columns, prioritizing those ending with "_PERIODE".}

\item{end_date_col}{The name of the column containing end dates.
Default is NULL, which means the function will automatically search for
appropriate columns, prioritizing those ending with "_PERIODE".}
}
\value{
A filtered tibble containing only records that are valid at the reference date.
}
\description{
This function filters temporal data based on a reference date,
keeping only records that are valid at the specified date.
It prioritizes columns ending with "_PERIODE" for date filtering.
}
\examples{
\dontrun{
# Get educational institutions valid on January 1, 2023
institutions <- rio_get_data(dataset_name = "onderwijsaanbieders")
current_institutions <- rio_filter_by_reference_date(institutions, as.Date("2023-01-01"))

# With custom column names
custom_data <- rio_get_data(dataset_name = "custom_dataset")
filtered_data <- rio_filter_by_reference_date(custom_data, as.Date("2023-01-01"),
                                              start_date_col = "STARTDATUM",
                                              end_date_col = "UITBEDRIJFDATUM")
}

}
