% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_find_connection_paths}
\alias{rio_find_connection_paths}
\title{Find connection paths between RIO datasets}
\usage{
rio_find_connection_paths(
  ...,
  structure = NULL,
  max_path_length = 3,
  min_confidence = "medium"
)
}
\arguments{
\item{...}{Named or unnamed character strings of dataset names to connect.}

\item{structure}{The RIO structure mapping as returned by rio_load_structure().
If NULL, it will be loaded automatically.}

\item{max_path_length}{Maximum length of connection paths to consider. Default is 3.}

\item{min_confidence}{Minimum confidence level for connections. Can be "high" or "medium".
Default is "medium".}
}
\value{
A list of possible connection paths between the datasets.
}
\description{
This function finds all possible connection paths between specified datasets
using the stored RIO structure mapping.
}
\examples{
\dontrun{
# Find connection paths between three datasets
paths <- rio_find_connection_paths("onderwijslocaties", "vestigingserkenningen", "onderwijsaanbieders")
}

}
