% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{rio_get_dataset_metadata}
\alias{rio_get_dataset_metadata}
\title{Get metadata of a dataset}
\usage{
rio_get_dataset_metadata(datasets, dataset_name)
}
\arguments{
\item{datasets}{The RIO datasets structure as returned by rio_load_datasets().}

\item{dataset_name}{Name of the dataset.}
}
\value{
A list with the metadata of the dataset.
}
\description{
This function retrieves the metadata of a specific dataset in the RIO structure.
}
\keyword{internal}
