% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{rio_get_metadata}
\alias{rio_get_metadata}
\title{Get metadata for the RIO dataset}
\usage{
rio_get_metadata()
}
\value{
A list containing the package metadata.
}
\description{
This function retrieves detailed metadata for the Dutch Register of Institutions
and Programs (RIO) dataset.
}
\examples{
\dontrun{
rio_metadata <- rio_get_metadata()
}

}
\keyword{internal}
