% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{rio_get_resource_info}
\alias{rio_get_resource_info}
\title{Get detailed information about a specific dataset}
\usage{
rio_get_resource_info(resource_id = NULL, resource_name = NULL)
}
\arguments{
\item{resource_id}{The ID of the resource to retrieve. Default is NULL.}

\item{resource_name}{The name of the dataset resource to retrieve. Default is NULL.
Either resource_id or resource_name must be provided.}
}
\value{
A list containing detailed information about the dataset.
}
\description{
This function retrieves detailed information about a specific dataset (resource) in the RIO package.
}
\keyword{internal}
