% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{rio_get_resource_info}
\alias{rio_get_resource_info}
\title{Get detailed information about a specific dataset}
\usage{
rio_get_resource_info(dataset_id = NULL, dataset_name = NULL)
}
\arguments{
\item{dataset_id}{The ID of the resource to retrieve. Default is NULL.}

\item{dataset_name}{The name of the dataset resource to retrieve. Default is NULL.
Either dataset_id or dataset_name must be provided.}
}
\value{
A list containing detailed information about the dataset.
}
\description{
This function retrieves detailed information about a specific dataset (resource) in the RIO package.
}
\keyword{internal}
