% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_join}
\alias{rio_join}
\title{Join RIO datasets with dataframe lookup}
\usage{
rio_join(df, id_col, resource_id, rio_id_field = NULL, reference_date = NULL)
}
\arguments{
\item{df}{A dataframe containing IDs to look up.}

\item{id_col}{The name of the column in the dataframe containing IDs.}

\item{resource_id}{The ID of the dataset resource to retrieve.}

\item{rio_id_field}{The name of the ID field in the RIO dataset. If NULL, uses id_col.}

\item{reference_date}{Optional date to filter valid records. Default is the current date.}
}
\value{
A tibble with the input dataframe joined with data from the RIO dataset.
}
\description{
This function joins data from a RIO dataset based on IDs in a dataframe.
}
\examples{
\dontrun{
# Join a dataframe with institution data
df <- data.frame(inst_id = c("12345", "67890"), value = c(10, 20))
result <- rio_join(
  df = df,
  id_col = "inst_id",
  resource_id = "institution-resource-id",
  rio_id_field = "INSTELLINGSCODE"
)
}

}
