% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{rio_list_datasets}
\alias{rio_list_datasets}
\title{List available datasets in the RIO package}
\usage{
rio_list_datasets(conn = NULL, package_id = "rio_nfo_po_vo_vavo_mbo_ho")
}
\arguments{
\item{conn}{A connection object created with \code{rio_api_connection()}.
If NULL, a new connection will be created.}

\item{package_id}{The ID of the package to retrieve datasets from. Default is "rio_nfo_po_vo_vavo_mbo_ho".}
}
\value{
A tibble with information about each dataset, including ID, name, and description.
}
\description{
This function retrieves a list of all available datasets (resources) in the RIO package.
}
\examples{
\dontrun{
datasets <- rio_list_datasets()
}

}
