% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations_utils.R
\name{rio_list_relations}
\alias{rio_list_relations}
\title{List all relations in the RIO structure}
\usage{
rio_list_relations(relations, as_dataframe = TRUE)
}
\arguments{
\item{relations}{The list of RIO relations.}

\item{as_dataframe}{Logical indicating whether the result should be returned as a dataframe.
Default is TRUE.}
}
\value{
A dataframe or list with all relations.
}
\description{
This function provides an overview of all defined relations.
}
\keyword{internal}
