% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{rio_list_tables}
\alias{rio_list_tables}
\title{List available tables in the RIO package}
\usage{
rio_list_tables()
}
\value{
A tibble with information about each table, including ID, name, and description.
}
\description{
This function retrieves a list of all available tables (resources) in the RIO package.
}
\examples{
\dontrun{
tables <- rio_list_tables()
}

}
