% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_load_structure}
\alias{rio_load_structure}
\title{Laad de gedetecteerde RIO structuur}
\usage{
rio_load_structure(auto_detect = TRUE, days_threshold = 28, quiet = FALSE)
}
\arguments{
\item{auto_detect}{Logische waarde die aangeeft of automatisch een nieuwe structuurdetectie
moet worden uitgevoerd als er geen bestaat of als deze verouderd is. Standaard is TRUE.}

\item{days_threshold}{Aantal dagen waarna een structuurdetectie als verouderd wordt beschouwd.
Standaard is 28 (4 weken).}

\item{quiet}{Logische waarde die aangeeft of voortgangsberichten onderdrukt moeten worden. Standaard is FALSE.}
}
\value{
Een lijst met de gedetecteerde RIO structuur.
}
\description{
Deze functie laadt de opgeslagen RIO structuur. Als er geen structuur bestaat of deze
verouderd is, kan optioneel een nieuwe detectie worden uitgevoerd.
}
\examples{
\dontrun{
# Laad de RIO structuur
structure <- rio_load_structure()
}

}
\keyword{internal}
