% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_lookup}
\alias{rio_lookup}
\title{Join RIO datasets with vector lookup}
\usage{
rio_lookup(ids, resource_id, id_field, reference_date = NULL)
}
\arguments{
\item{ids}{A vector of IDs to look up.}

\item{resource_id}{The ID of the dataset resource to retrieve.}

\item{id_field}{The name of the ID field in the dataset.}

\item{reference_date}{Optional date to filter valid records. Default is the current date.}
}
\value{
A tibble with data for the specified IDs.
}
\description{
This function joins data from a RIO dataset based on a vector of IDs.
}
\keyword{internal}
