% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{rio_visualize_connections}
\alias{rio_visualize_connections}
\title{Visualiseer gedetecteerde verbindingen tussen datasets}
\usage{
rio_visualize_connections(
  connections,
  min_confidence = "medium",
  interactive = TRUE
)
}
\arguments{
\item{connections}{Een lijst met verbindingen zoals geretourneerd door rio_auto_connect()}

\item{min_confidence}{Minimaal betrouwbaarheidsniveau voor verbindingen}

\item{interactive}{Logische waarde die aangeeft of een interactieve visualisatie moet worden gemaakt}
}
\value{
Een visualisatie-object
}
\description{
Visualiseer gedetecteerde verbindingen tussen datasets
}
\keyword{internal}
